+++ src/features/scoreManagementV2/components/InputWinPoint/index.tsx
